[API] Rename campaignParticipationResult folder
In campaign-participation-result according to others
