+++ api/lib/application/campaign-participation-results/campaign-participation-result-controller.js
