+++ api/lib/application/campaign-participation-results/index.js
